+++ packages/yarnpkg-nm/sources/buildNodeModulesTree.ts
@@ -53,6 +53,7 @@ export interface NodeModulesTreeOptions {
   pnpifyFs?: boolean;
   validateExternalSoftLinks?: boolean;
   hoistingLimitsByCwd?: Map<PortablePath, NodeModulesHoistingLimits>;
+  selfReferencesByCwd?: Map<PortablePath, Boolean>;
   project?: Project;
 }
 
@@ -346,7 +347,8 @@ const buildPackageTree = (pnp: PnpApi, options: NodeModulesTreeOptions): { packa
       }
     }
 
-    parent.dependencies.add(node);
+    if (pkg !== parentPkg || pkg.linkType !== LinkType.SOFT || !options.selfReferencesByCwd || options.selfReferencesByCwd.get(parentRelativeCwd))
+      parent.dependencies.add(node);
 
     const isWorkspaceDependency = locator !== topLocator && pkg.linkType === LinkType.SOFT && !locator.name.endsWith(WORKSPACE_NAME_SUFFIX) && !isExternalSoftLinkPackage;